src/ml-services/requirements.txt
flask
flask-sqlalchemy
huggingface_hub
flask-login
werkzeug
tensorflow==2.16.1
numpy
opencv-python
pillow
email-validator==1.1.3
python-dotenv==0.19.0 
fastapi
uvicorn[standard]
pillow
tensorflow-cpu